Refactor cargo_heather: stream-based public API + fixture-driven tests
- Replace path-based public API with stream-based check<R: Read> /
  fix<R: Read, W: Write> taking impl Read/impl Write
- Move bin-only modules (cli, config, scanner, run) out of the library
  into src/bin/cargo-heather/, leaving the library as pure transformation
  primitives (no FS, no CLI, no config parsing)
- Replace inline mod tests blocks and tests/integration_test.rs with
  fixture-driven integration tests in tests/cases.rs + tests/fixtures/
  (rust_ok/missing/mismatch/descriptive/empty, toml_ok/missing/mismatch,
  script_ok/missing/mismatch)
- Drop assert_cmd, predicates, tempfile dev-deps
- Set allowed_external_types = [] (no external types in lib public API)
