+++ crates/cargo_heather/Cargo.toml
@@ -20,11 +20,11 @@ repository = "https://github.com/microsoft/ox-tools/tree/main/crates/cargo_heath
 all-features = true
 
 [package.metadata.cargo_check_external_types]
-allowed_external_types = ["clap_builder::derive::*", "ohno::app::error::AppError"]
+allowed_external_types = []
 
 [[bin]]
 name = "cargo-heather"
-path = "src/main.rs"
+path = "src/bin/cargo-heather/main.rs"
 
 [dependencies]
 clap = { workspace = true, features = ["derive", "std", "help", "usage", "error-context"] }
@@ -34,11 +34,6 @@ toml = { workspace = true, features = ["parse", "display"] }
 tracing = { workspace = true }
 tracing-subscriber = { workspace = true, features = ["fmt"] }
 walkdir = { workspace = true }
-
-[dev-dependencies]
-assert_cmd = { workspace = true }
-predicates = { workspace = true }
-tempfile = { workspace = true }
 
 [lints]
 workspace = true
