+++ crates/cargo_heather/src/checker/extract.rs
@@ -55,55 +55,3 @@ pub(super) fn script_header(content: &str, style: CommentStyle) -> Option<String
     }
     collect_comment_block(lines, style)
 }
-
-#[cfg(test)]
-#[cfg_attr(coverage_nightly, coverage(off))]
-mod tests {
-    use super::*;
-
-    #[test]
-    fn extract_header_simple() {
-        let content = "// Hello\n// World\n\nfn main() {}\n";
-        let header = header_comment(content, CommentStyle::DoubleSlash).unwrap();
-        assert_eq!(header, "Hello\nWorld");
-    }
-
-    #[test]
-    fn extract_header_with_blank_comment_line() {
-        let content = "// First\n//\n// Third\n\nfn main() {}\n";
-        let header = header_comment(content, CommentStyle::DoubleSlash).unwrap();
-        assert_eq!(header, "First\n\nThird");
-    }
-
-    #[test]
-    fn extract_no_header() {
-        let content = "fn main() {}\n";
-        assert!(header_comment(content, CommentStyle::DoubleSlash).is_none());
-    }
-
-    #[test]
-    fn extract_empty_content() {
-        assert!(header_comment("", CommentStyle::DoubleSlash).is_none());
-    }
-
-    #[test]
-    fn extract_header_trailing_empty_comment_lines_trimmed() {
-        let content = "// Licensed under MIT\n//\n\nfn main() {}\n";
-        let extracted = header_comment(content, CommentStyle::DoubleSlash);
-        assert_eq!(extracted.as_deref(), Some("Licensed under MIT"));
-    }
-
-    #[test]
-    fn extract_script_header_trailing_empty_comment_lines_trimmed() {
-        let content = "#!/usr/bin/env -S cargo +nightly -Zscript\n---\n# Licensed under MIT\n#\n---\n";
-        let extracted = script_header(content, CommentStyle::Hash);
-        assert_eq!(extracted.as_deref(), Some("Licensed under MIT"));
-    }
-
-    #[test]
-    fn extract_script_header_no_dash_line() {
-        let content = "#!/usr/bin/env cargo\nnot-a-dash\n# License\n";
-        let extracted = script_header(content, CommentStyle::Hash);
-        assert!(extracted.is_none());
-    }
-}
