+++ crates/cargo_heather/src/checker/matcher.rs
@@ -31,60 +31,3 @@ pub(super) fn normalize_to_lines(text: &str) -> Vec<String> {
     }
     lines
 }
-
-#[cfg(test)]
-#[cfg_attr(coverage_nightly, coverage(off))]
-mod tests {
-    use super::*;
-
-    #[test]
-    fn normalize_to_lines_trims_trailing_whitespace() {
-        let result = normalize_to_lines("Hello   \nWorld  ");
-        assert_eq!(result, vec!["Hello".to_owned(), "World".to_owned()]);
-    }
-
-    #[test]
-    fn normalize_to_lines_strips_outer_blank_lines() {
-        let result = normalize_to_lines("\n\nHello\nWorld\n\n");
-        assert_eq!(result, vec!["Hello".to_owned(), "World".to_owned()]);
-    }
-
-    #[test]
-    fn header_matches_exact_single_line() {
-        assert!(header_matches("MIT License", "MIT License"));
-    }
-
-    #[test]
-    fn header_matches_exact_multi_line() {
-        assert!(header_matches(
-            "Copyright Microsoft.\nMIT License.",
-            "Copyright Microsoft.\nMIT License."
-        ));
-    }
-
-    #[test]
-    fn header_matches_with_extra_trailing_lines() {
-        assert!(header_matches(
-            "Copyright Microsoft.\nMIT License.\nDescriptive comment.",
-            "Copyright Microsoft.\nMIT License."
-        ));
-    }
-
-    #[test]
-    fn header_matches_rejects_when_first_line_differs() {
-        assert!(!header_matches(
-            "Apache License.\nMIT License.",
-            "Copyright Microsoft.\nMIT License."
-        ));
-    }
-
-    #[test]
-    fn header_matches_rejects_when_actual_too_short() {
-        assert!(!header_matches("Copyright Microsoft.", "Copyright Microsoft.\nMIT License."));
-    }
-
-    #[test]
-    fn header_matches_tolerates_trailing_whitespace_on_lines() {
-        assert!(header_matches("MIT License   ", "MIT License"));
-    }
-}
