+++ crates/cargo_heather/src/comment.rs
@@ -154,200 +154,3 @@ impl CommentStyle {
         }
     }
 }
-
-#[cfg(test)]
-#[cfg_attr(coverage_nightly, coverage(off))]
-mod tests {
-    use std::path::PathBuf;
-
-    use super::*;
-
-    #[test]
-    fn from_path_rs() {
-        assert_eq!(
-            CommentStyle::from_path(&PathBuf::from("src/main.rs")),
-            Some(CommentStyle::DoubleSlash)
-        );
-    }
-
-    #[test]
-    fn from_path_toml() {
-        assert_eq!(CommentStyle::from_path(&PathBuf::from("Cargo.toml")), Some(CommentStyle::Hash));
-    }
-
-    #[test]
-    fn from_path_unsupported() {
-        assert_eq!(CommentStyle::from_path(&PathBuf::from("README.md")), None);
-    }
-
-    #[test]
-    fn from_path_no_extension() {
-        assert_eq!(CommentStyle::from_path(&PathBuf::from("Makefile")), None);
-    }
-
-    #[test]
-    fn prefix_double_slash() {
-        assert_eq!(CommentStyle::DoubleSlash.prefix(), "//");
-        assert_eq!(CommentStyle::DoubleSlash.prefix_space(), "// ");
-    }
-
-    #[test]
-    fn prefix_hash() {
-        assert_eq!(CommentStyle::Hash.prefix(), "#");
-        assert_eq!(CommentStyle::Hash.prefix_space(), "# ");
-    }
-
-    #[test]
-    fn format_header_double_slash_single_line() {
-        let result = CommentStyle::DoubleSlash.format_header("MIT License");
-        assert_eq!(result, "// MIT License");
-    }
-
-    #[test]
-    fn format_header_double_slash_multiline() {
-        let result = CommentStyle::DoubleSlash.format_header("Line one\n\nLine three");
-        assert_eq!(result, "// Line one\n//\n// Line three");
-    }
-
-    #[test]
-    fn format_header_hash_single_line() {
-        let result = CommentStyle::Hash.format_header("MIT License");
-        assert_eq!(result, "# MIT License");
-    }
-
-    #[test]
-    fn format_header_hash_multiline() {
-        let result = CommentStyle::Hash.format_header("Line one\n\nLine three");
-        assert_eq!(result, "# Line one\n#\n# Line three");
-    }
-
-    #[test]
-    fn is_header_comment_line_rs_regular() {
-        let style = CommentStyle::DoubleSlash;
-        assert!(style.is_header_comment_line("// hello"));
-        assert!(style.is_header_comment_line("//"));
-        assert!(style.is_header_comment_line("// "));
-    }
-
-    #[test]
-    fn is_header_comment_line_rs_excludes_doc() {
-        let style = CommentStyle::DoubleSlash;
-        assert!(!style.is_header_comment_line("/// doc"));
-        assert!(!style.is_header_comment_line("//! module doc"));
-    }
-
-    #[test]
-    fn is_header_comment_line_rs_rejects_non_comment() {
-        let style = CommentStyle::DoubleSlash;
-        assert!(!style.is_header_comment_line("fn main() {}"));
-        assert!(!style.is_header_comment_line(""));
-    }
-
-    #[test]
-    fn is_header_comment_line_hash() {
-        let style = CommentStyle::Hash;
-        assert!(style.is_header_comment_line("# hello"));
-        assert!(style.is_header_comment_line("#"));
-        assert!(style.is_header_comment_line("# "));
-    }
-
-    #[test]
-    fn is_header_comment_line_hash_rejects_non_comment() {
-        let style = CommentStyle::Hash;
-        assert!(!style.is_header_comment_line("[package]"));
-        assert!(!style.is_header_comment_line(""));
-        assert!(!style.is_header_comment_line("name = \"foo\""));
-    }
-
-    #[test]
-    fn strip_prefix_double_slash_with_space() {
-        assert_eq!(CommentStyle::DoubleSlash.strip_prefix("// Hello"), "Hello");
-    }
-
-    #[test]
-    fn strip_prefix_double_slash_without_space() {
-        assert_eq!(CommentStyle::DoubleSlash.strip_prefix("//Hello"), "Hello");
-    }
-
-    #[test]
-    fn strip_prefix_double_slash_empty() {
-        assert_eq!(CommentStyle::DoubleSlash.strip_prefix("//"), "");
-    }
-
-    #[test]
-    fn strip_prefix_hash_with_space() {
-        assert_eq!(CommentStyle::Hash.strip_prefix("# Hello"), "Hello");
-    }
-
-    #[test]
-    fn strip_prefix_hash_without_space() {
-        assert_eq!(CommentStyle::Hash.strip_prefix("#Hello"), "Hello");
-    }
-
-    #[test]
-    fn strip_prefix_hash_empty() {
-        assert_eq!(CommentStyle::Hash.strip_prefix("#"), "");
-    }
-
-    #[test]
-    fn is_cargo_script_detects_shebang_and_frontmatter() {
-        let content = "#!/usr/bin/env -S cargo +nightly -Zscript\n---\n# License\n";
-        assert!(is_cargo_script(content));
-    }
-
-    #[test]
-    fn is_cargo_script_rejects_inner_attribute() {
-        let content = "#![allow(unused)]\nfn main() {}\n";
-        assert!(!is_cargo_script(content));
-    }
-
-    #[test]
-    fn is_cargo_script_rejects_no_frontmatter() {
-        let content = "#!/usr/bin/env cargo\nfn main() {}\n";
-        assert!(!is_cargo_script(content));
-    }
-
-    #[test]
-    fn is_cargo_script_rejects_empty() {
-        assert!(!is_cargo_script(""));
-    }
-
-    #[test]
-    fn is_cargo_script_rejects_non_shebang_with_frontmatter() {
-        // First line doesn't start with `#!` but second line is `---`
-        let content = "fn main() {}\n---\nsome content\n";
-        assert!(!is_cargo_script(content));
-    }
-
-    #[test]
-    fn file_kind_detect_regular_rs() {
-        let kind = FileKind::detect(&PathBuf::from("src/main.rs"), Some("fn main() {}"));
-        assert_eq!(kind, Some(FileKind::Rust));
-    }
-
-    #[test]
-    fn file_kind_detect_cargo_script() {
-        let content = "#!/usr/bin/env -S cargo +nightly -Zscript\n---\n";
-        let kind = FileKind::detect(&PathBuf::from("script.rs"), Some(content));
-        assert_eq!(kind, Some(FileKind::CargoScript));
-    }
-
-    #[test]
-    fn file_kind_detect_toml() {
-        let kind = FileKind::detect(&PathBuf::from("Cargo.toml"), None);
-        assert_eq!(kind, Some(FileKind::Toml));
-    }
-
-    #[test]
-    fn file_kind_detect_unsupported() {
-        let kind = FileKind::detect(&PathBuf::from("README.md"), None);
-        assert_eq!(kind, None);
-    }
-
-    #[test]
-    fn file_kind_comment_style() {
-        assert_eq!(FileKind::Rust.comment_style(), CommentStyle::DoubleSlash);
-        assert_eq!(FileKind::Toml.comment_style(), CommentStyle::Hash);
-        assert_eq!(FileKind::CargoScript.comment_style(), CommentStyle::Hash);
-    }
-}
