+++ crates/cargo_heather/src/error.rs
@@ -74,91 +74,3 @@ impl std::error::Error for HeatherError {
         }
     }
 }
-
-#[cfg(test)]
-#[cfg_attr(coverage_nightly, coverage(off))]
-mod tests {
-    use super::*;
-
-    #[test]
-    fn file_read_error_display() {
-        let err = HeatherError::FileRead {
-            path: PathBuf::from("src/main.rs"),
-            source: std::io::Error::new(std::io::ErrorKind::NotFound, "not found"),
-        };
-        let msg = err.to_string();
-        assert!(msg.contains("src/main.rs"));
-        assert!(msg.contains("not found"));
-    }
-
-    #[test]
-    fn config_parse_error_display() {
-        let err = HeatherError::ConfigParse {
-            path: PathBuf::from(".cargo-heather.toml"),
-            message: "bad syntax".into(),
-        };
-        assert!(err.to_string().contains("bad syntax"));
-    }
-
-    #[test]
-    fn config_not_found_error_display() {
-        let err = HeatherError::ConfigNotFound(PathBuf::from(".cargo-heather.toml"));
-        assert!(err.to_string().contains(".cargo-heather.toml"));
-    }
-
-    #[test]
-    fn config_invalid_error_display() {
-        let err = HeatherError::ConfigInvalid("both license and header specified".into());
-        assert!(err.to_string().contains("both license and header"));
-    }
-
-    #[test]
-    fn unknown_license_error_display() {
-        let err = HeatherError::UnknownLicense("FAKE-1.0".into());
-        assert!(err.to_string().contains("FAKE-1.0"));
-    }
-
-    #[test]
-    fn validation_failed_error_display() {
-        let err = HeatherError::ValidationFailed(3);
-        assert!(err.to_string().contains("3 file(s)"));
-    }
-
-    #[test]
-    fn unsupported_file_type_error_display() {
-        let err = HeatherError::UnsupportedFileType {
-            path: PathBuf::from("data.csv"),
-        };
-        assert!(err.to_string().contains("unsupported file type"));
-        assert!(err.to_string().contains("data.csv"));
-    }
-
-    #[test]
-    fn error_source_file_read() {
-        use std::error::Error;
-        let err = HeatherError::FileRead {
-            path: PathBuf::from("missing.rs"),
-            source: std::io::Error::new(std::io::ErrorKind::NotFound, "gone"),
-        };
-        assert!(err.source().is_some());
-    }
-
-    #[test]
-    fn error_source_none_for_other_variants() {
-        use std::error::Error;
-        let variants: Vec<HeatherError> = vec![
-            HeatherError::ConfigNotFound(PathBuf::from("x")),
-            HeatherError::ConfigInvalid("bad".into()),
-            HeatherError::UnknownLicense("X".into()),
-            HeatherError::UnsupportedFileType { path: PathBuf::from("x") },
-            HeatherError::ValidationFailed(1),
-            HeatherError::ConfigParse {
-                path: PathBuf::from("x"),
-                message: "y".into(),
-            },
-        ];
-        for v in &variants {
-            assert!(v.source().is_none());
-        }
-    }
-}
