+++ crates/cargo_heather/src/license.rs
@@ -248,88 +248,3 @@ pub fn supported_licenses() -> Vec<&'static str> {
     ids.sort_unstable();
     ids
 }
-
-#[cfg(test)]
-#[cfg_attr(coverage_nightly, coverage(off))]
-mod tests {
-    use super::*;
-
-    #[test]
-    fn lookup_mit() {
-        let header = header_for_license("MIT").unwrap();
-        assert_eq!(header, "Licensed under the MIT License.");
-    }
-
-    #[test]
-    fn lookup_apache() {
-        let header = header_for_license("Apache-2.0").unwrap();
-        assert!(header.contains("Apache License, Version 2.0"));
-        assert!(header.contains("http://www.apache.org/licenses/LICENSE-2.0"));
-    }
-
-    #[test]
-    fn lookup_case_insensitive() {
-        let header = header_for_license("mit").unwrap();
-        assert_eq!(header, "Licensed under the MIT License.");
-
-        let header2 = header_for_license("MPL-2.0").unwrap();
-        assert!(header2.contains("Mozilla Public"));
-    }
-
-    #[test]
-    fn lookup_unknown_license() {
-        let err = header_for_license("FAKE-LICENSE-1.0").unwrap_err();
-        assert!(err.to_string().contains("FAKE-LICENSE-1.0"));
-    }
-
-    #[test]
-    fn all_licenses_have_nonempty_headers() {
-        for lic in LICENSES {
-            assert!(!lic.header.is_empty(), "License {} has an empty header", lic.id);
-        }
-    }
-
-    #[test]
-    fn supported_licenses_returns_sorted() {
-        let ids = supported_licenses();
-        assert!(ids.len() >= 15);
-
-        let mut sorted = ids.clone();
-        sorted.sort_unstable();
-        assert_eq!(ids, sorted);
-    }
-
-    #[test]
-    fn all_gpl_variants_present() {
-        for id in &["GPL-2.0-only", "GPL-2.0-or-later", "GPL-3.0-only", "GPL-3.0-or-later"] {
-            assert!(header_for_license(id).is_ok(), "Missing license: {id}");
-        }
-    }
-
-    #[test]
-    fn all_lgpl_variants_present() {
-        for id in &["LGPL-2.1-only", "LGPL-2.1-or-later", "LGPL-3.0-only", "LGPL-3.0-or-later"] {
-            assert!(header_for_license(id).is_ok(), "Missing license: {id}");
-        }
-    }
-
-    #[test]
-    fn bsd_licenses_present() {
-        header_for_license("BSD-2-Clause").unwrap();
-        header_for_license("BSD-3-Clause").unwrap();
-        header_for_license("0BSD").unwrap();
-    }
-
-    #[test]
-    fn misc_licenses_present() {
-        for id in &["ISC", "Unlicense", "BSL-1.0", "Zlib"] {
-            header_for_license(id).unwrap();
-        }
-    }
-
-    #[test]
-    fn agpl_licenses_present() {
-        header_for_license("AGPL-3.0-only").unwrap();
-        header_for_license("AGPL-3.0-or-later").unwrap();
-    }
-}
