+++ crates/cargo_heather/tests/fixtures/rust_empty/expected.rs
@@ -1,0 +1,1 @@
+// Licensed under the MIT License.
