+++ crates/cargo_heather/tests/fixtures/rust_mismatch/input.rs
@@ -1,0 +1,3 @@
+// Licensed under the Apache License.
+
+fn main() {}
