+++ crates/cargo_heather/tests/fixtures/rust_missing/expected.rs
@@ -1,0 +1,3 @@
+// Licensed under the MIT License.
+
+fn main() {}
