+++ crates/cargo_heather/tests/fixtures/rust_missing/input.rs
@@ -1,0 +1,1 @@
+fn main() {}
