+++ crates/cargo_heather/tests/fixtures/script_mismatch/input.rs
@@ -1,0 +1,5 @@
+#!/usr/bin/env cargo
+---
+# Licensed under the Apache License.
+
+fn main() {}
