+++ crates/cargo_heather/tests/fixtures/script_missing/expected.rs
@@ -1,0 +1,5 @@
+#!/usr/bin/env cargo
+---
+# Licensed under the MIT License.
+
+fn main() {}
