+++ crates/cargo_heather/tests/fixtures/script_missing/input.rs
@@ -1,0 +1,3 @@
+#!/usr/bin/env cargo
+---
+fn main() {}
