+++ crates/cargo_heather/tests/fixtures/toml_mismatch/input.toml
@@ -1,0 +1,4 @@
+# Licensed under the Apache License.
+
+[package]
+name = "foo"
