+++ crates/cargo_heather/tests/fixtures/toml_missing/input.toml
@@ -1,0 +1,2 @@
+[package]
+name = "foo"
