+++ crates/cargo_heather/tests/fixtures/toml_ok/expected.toml
@@ -1,0 +1,4 @@
+# Licensed under the MIT License.
+
+[package]
+name = "foo"
